# Slightly speed up `bundler/setup`