--- a/bundler/lib/bundler/spec_set.rb
+++ b/bundler/lib/bundler/spec_set.rb
@@ -18,7 +18,7 @@ module Bundler
 
       loop do
         break unless dep = deps.shift
-        next if handled.any?{|d| d.name == dep.name && (match_current_platform || d.__platform == dep.__platform) }
+        next if handled.any?{|d| d.name == dep.name && (match_current_platform || d.__platform == dep.__platform) } || dep.name == "bundler"
 
         handled << dep
 
@@ -42,7 +42,7 @@ module Bundler
       end
 
       if spec = lookup["bundler"].first
-        specs |= [spec]
+        specs << spec
       end
 
       check ? true : specs